+++ CHANGELOG.md
@@ -14,10 +14,10 @@
 - Fixed ONTAP SAN drivers to clean up if provisioning fails (Issue [#442](https://github.com/NetApp/trident/issues/442)).
 
 **Enhancements:**
-- **Kubernetes:** Added automatic node preparation for NFS protocol. By default, Trident will now attempt to make
- sure NFS packages and services are installed and running the first time an NFS volume is mounted on a worker node.
-  Can be disabled with `--disable-node-prep` installation option.
 - Added `qtreesPerFlexvol` option to allow customizing the number of qtrees per FlexVol in the ontap-nas-economy driver.
+
+**Tech Previews:**
+- **Kubernetes:** Automatic node preparation for NFS and iSCSI protocols. Trident can now attempt to make sure NFS and/or iSCSI packages and services are installed and running the first time an NFS or iSCSI volume is mounted on a worker node. Can be enabled with `--enable-node-prep` installation option.
 
 **Deprecations:**
 - **Kubernetes:** Removed etcd support, including ability to upgrade directly from etcd-based Trident (v19.04.1 or older).